# Analyze and optimize test performance