--- a/OPTIMIZATION.md
+++ b/OPTIMIZATION.md
@@ -1,0 +1,90 @@
+# FastLED Test Suite Optimization Analysis
+
+## Current Performance Baseline
+
+**Total Test Execution Time: ~152 seconds**
+- Compilation Phase: ~5.5 seconds
+- Test Execution Phase: ~4.5 seconds  
+- Overhead/Infrastructure: ~142 seconds
+
+## Analysis Summary
+
+From examining the test output and codebase, the primary bottlenecks are:
+
+### 1. **Serial Test Execution** (Primary Issue)
+- Tests are run with `--no-parallel` flag, forcing sequential execution
+- 89 unit tests each taking individual process spawn/teardown cycles
+- Estimated savings: **60-80% reduction** in total time
+
+### 2. **Process Spawn Overhead**
+- Each test spawns individual processes via Python subprocess
+- Heavy use of `uv run python` wrapper adds ~100-200ms per test
+- 89 tests × 150ms overhead = ~13 seconds of pure spawn overhead
+
+### 3. **Duplicate Test Execution**
+- Tests appear to run twice in some configurations (noticed duplicate output)
+- Some tests like `test_fft` show identical repeated output blocks
+
+### 4. **Memory Buffer Inefficiencies**
+- `OutputBuffer` class uses complex queuing system for simple sequential output
+- Thread synchronization overhead for ordering that's unnecessary in single-threaded cases
+
+### 5. **Redundant File System Operations**
+- Multiple checks for test file changes
+- Repeated file discovery operations
+- Cache directory operations that could be optimized
+
+## Specific Optimization Opportunities
+
+### A. **Enable Parallel Test Execution** 
+**Impact: HIGH (60-80% time reduction)**
+- Remove `--no-parallel` default
+- Implement proper parallel test runner
+- Use CPU count-based parallelism (detected: available CPUs)
+
+### B. **Optimize Process Spawning**
+**Impact: MEDIUM (10-15% time reduction)**
+- Replace `uv run python` with direct Python executable calls where possible
+- Use `sys.executable` directly to avoid shell resolution
+- Implement test batching to reduce process spawn count
+
+### C. **Eliminate Duplicate Test Runs**
+**Impact: MEDIUM (potential 50% reduction if confirmed)**
+- Investigate why some tests appear to execute twice
+- Consolidate test execution paths
+
+### D. **Streamline Output Handling**
+**Impact: LOW-MEDIUM (5-10% time reduction)**
+- Replace complex `OutputBuffer` with simple sequential output in non-parallel mode
+- Reduce Unicode/encoding overhead
+- Minimize flush operations
+
+### E. **Optimize File System Operations**
+**Impact: LOW (2-5% time reduction)**
+- Cache test file discovery results
+- Reduce redundant file existence checks
+- Optimize build directory operations
+
+### F. **Unity Builds Optimization**
+**Impact: MEDIUM (compilation time only)**
+- Current tests disable unity builds (`--no-unity`)
+- Enable unity builds for faster compilation when appropriate
+
+## Implementation Priority
+
+1. **Enable Parallel Testing** - Biggest impact, moderate effort
+2. **Fix Duplicate Execution** - High impact if confirmed, low effort  
+3. **Process Spawn Optimization** - Medium impact, low effort
+4. **Output Stream Optimization** - Medium impact, low effort
+5. **Unity Build Enablement** - Medium impact, configuration change
+6. **File System Optimizations** - Low impact, low effort
+
+## Expected Results
+
+**Conservative Estimate:**
+- From 152s to ~45-60s (60-70% improvement)
+
+**Optimistic Estimate:** 
+- From 152s to ~25-35s (75-85% improvement)
+
+The most significant gains will come from parallel execution, which is currently disabled but the infrastructure exists to support it.